pokermate/source/requirements.txt
Pillow
ultralytics==8.0.208
treys
numpy
opencv-python-headless
streamlit